requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
boto3==1.35.51
botocore==1.35.51
cachetools==5.5.0
catalogue==2.0.10
catboost==1.1.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
datasets==2.14.4
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
dgl==2.2.0
dill==0.3.7
distlib==0.3.9
docker-pycreds==0.4.0
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
fastai==2.7.18
fastcore==1.7.19
fastdownload==0.0.7
fastjsonschema==2.20.0
fastprogress==1.0.3
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gluonts==0.15.1
google-api-core==2.22.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
graphviz==0.20.3
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyperopt==0.2.7
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lab==8.3
langcodes==3.4.1
language_data==1.2.0
lightgbm==4.3.0
lightning==2.3.3
lightning-utilities==0.11.8
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mlforecast==0.10.0
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nlpaug==1.1.11
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
nptyping==2.4.1
numba==0.60.0
numpy==1.26.4
nvidia-ml-py3==7.352.0
omegaconf==2.2.3
onnx==1.17.0
onnxruntime==1.19.2
opencensus==0.11.4
opencensus-context==0.1.3
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.0.11
optimum==1.17.1
ordered-set==4.1.0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.3.3
pytorch-metric-learning==2.3.0
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.19.3
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.18.0
seqeval==1.2.2
setproctitle==1.3.4
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
statsforecast==1.4.0
statsmodels==0.14.4
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==0.9.16
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.0.2
toolz==0.12.1
torch==2.3.0
torch-geometric==2.6.1
torch_cluster==1.6.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torch_spline_conv==1.2.2
torchaudio==2.3.0
torchdata==0.9.0
torchmetrics==1.2.1
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.40.2
txt2tags==3.9
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.20
utilsforecast==0.0.10
uv==0.4.28
virtualenv==20.27.1
wandb==0.18.7
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
window_ops==0.0.15
wrapt==1.16.0
xgboost==2.0.3
xxhash==3.5.0
yarl==1.17.0
zipp==3.20.2
